added in onClick for homeActivity adapter to go to match detail

+++ Game-Stats-for-Dota2/app/src/main/java/com/freatnor/game_stats_for_dota2/interfaces/MatchCallback.java
@@ -4,5 +4,5 @@ package com.freatnor.game_stats_for_dota2.interfaces;
  * Created by Jonathan Taylor on 9/7/16.
  */
 public interface MatchCallback {
-    void matchSelected(int matchId);
+    void matchSelected(long matchId);
 }
