+++ Game-Stats-for-Dota2/app/src/main/java/com/freatnor/game_stats_for_dota2/interfaces/MatchPlayerCallback.java
@@ -4,5 +4,5 @@ package com.freatnor.game_stats_for_dota2.interfaces;
  * Created by Jonathan Taylor on 9/7/16.
  */
 public interface MatchPlayerCallback {
-    void playerSelected(int playerId, int matchId);
+    void playerSelected(int playerId, long matchId);
 }
